gateway/requirements.txt
pyserial
paho-mqtt
influxdb-client
